% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapQTL.R
\name{BootstrapEQTL}
\alias{BootstrapEQTL}
\title{Bootstrap eQTL analysis for accurate effect size estimation}
\usage{
BootstrapEQTL(snps, gene, cvrt = SlicedData$new(), snpspos, genespos,
  n_bootstraps = 200, n_cores = 1)
}
\arguments{
\item{snps}{see MatrixEQTL}

\item{gene}{see MatrixEQTL}

\item{cvrt}{see MatrixEQTL}

\item{n_bootstraps}{number of bootstraps to run}

\item{n_cores}{number of cores to parallise the bootstrap procedure
over}

\item{...}{arguments for MatrixEQTL}
}
\description{
Performs bootstrap analysis on significant eQTL eGenes to obtain
accurate eQTL effect size estimates by correcting for the "Winner's
Curse".
}
\examples{
# Locations for example data from the MatrixEQTL package
base.dir = find.package('MatrixEQTL');
SNP_file_name = paste(base.dir, "/data/SNP.txt", sep="");
snps_location_file_name = paste(base.dir, "/data/snpsloc.txt", sep="");
expression_file_name = paste(base.dir, "/data/GE.txt", sep="");
gene_location_file_name = paste(base.dir, "/data/geneloc.txt", sep="");
covariates_file_name = paste(base.dir, "/data/Covariates.txt", sep="");

# Load the SNP data
snps = SlicedData$new();
snps$fileDelimiter = "\\t";      # the TAB character
snps$fileOmitCharacters = "NA"; # denote missing values;
snps$fileSkipRows = 1;          # one row of column labels
snps$fileSkipColumns = 1;       # one column of row labels
snps$fileSliceSize = 2000;      # read file in slices of 2,000 rows
snps$LoadFile(SNP_file_name);

# Load the data detailing the position of each SNP
snpspos = read.table(snps_location_file_name, header = TRUE, stringsAsFactors = FALSE);

# Load the gene expression data
gene = SlicedData$new();
gene$fileDelimiter = "\\t";      # the TAB character
gene$fileOmitCharacters = "NA"; # denote missing values;
gene$fileSkipRows = 1;          # one row of column labels
gene$fileSkipColumns = 1;       # one column of row labels
gene$fileSliceSize = 2000;      # read file in slices of 2,000 rows
gene$LoadFile(expression_file_name);

# Load the data detailing the position of each gene
genepos = read.table(gene_location_file_name, header = TRUE, stringsAsFactors = FALSE);

# Load the covariates data
cvrt = SlicedData$new();
cvrt$fileDelimiter = "\\t";      # the TAB character
cvrt$fileOmitCharacters = "NA"; # denote missing values;
cvrt$fileSkipRows = 1;          # one row of column labels
cvrt$fileSkipColumns = 1;       # one column of row labels
if(length(covariates_file_name)>0) {
  cvrt$LoadFile(covariates_file_name);
}

# Run the BootstrapEQTL analysis
eGenes <- BootstrapEQTL(snps, gene, cvrt, snpspos, genespos)

}
