% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapQTL.R
\name{BootstrapQTL}
\alias{BootstrapQTL}
\title{Bootstrap QTL analysis for accurate effect size estimation}
\usage{
BootstrapQTL(snps, gene, snpspos, genepos, cvrt = SlicedData$new(),
  n_bootstraps = 200, n_cores = 1, eGene_detection_file_name = NULL,
  bootstrap_file_directory = NULL, cisDist = 1e+06,
  local_correction = "bonferroni", global_correction = "fdr",
  bootstrap_eSNPs = "discovery", correction_type = "shrinkage",
  collate_top_snps = ifelse(bootstrap_eSNPs == "top", TRUE, FALSE),
  errorCovariance = numeric(), useModel = modelLINEAR)
}
\arguments{
\item{snps}{\code{\link[MatrixEQTL]{SlicedData}} object containing genotype
information used as input into \code{\link[MatrixEQTL]{Matrix_eQTL_main}}.}

\item{gene}{\code{\link[MatrixEQTL]{SlicedData}} object containing gene expression
information used as input into \code{\link[MatrixEQTL]{Matrix_eQTL_main}}.}

\item{snpspos}{\code{data.frame} object with information about SNP locations.
Used in conjunction with \code{'genespos'} and \code{'cisDist'} to
determine SNPs in \emph{cis} of each gene. Must have three columns: \enumerate{
  \item 'snpid' describing the name of the SNP and corresponding to rows in
  the 'snps' matrix.
  \item 'chr' describing the chromosome for each SNP.
  \item 'pos' describing the position of the SNP on the chromosome.
}}

\item{genepos}{\code{data.frame} object with information about transcript locations.
Used in conjunction with \code{'snpspos'} and \code{'cisDist'} to
determine SNPs in \emph{cis} of each gene. Must have four columns: \enumerate{
  \item 'geneid' describing the name of the gene and corresponding to rows in
  the 'gene' matrix.
  \item 'chr' describing the chromosome for each SNP.
  \item 'left' describing the start position of the transcript.
  \item 'right' describing the end position of the transcript.
}}

\item{cvrt}{\code{\link[MatrixEQTL]{SlicedData}} object containing covariate
information used as input into \code{\link[MatrixEQTL]{Matrix_eQTL_main}}.
Argument can be ignored in the case of no covariates.}

\item{n_bootstraps}{number of bootstraps to run.}

\item{n_cores}{number of cores to parallise the bootstrap procedure
over.}

\item{eGene_detection_file_name}{\code{character}, \code{connection} or \code{NULL}.
File to save local \emph{cis} associations to in the eGene detection analysis. Corresponds
to the \code{output_file_name.cis} argument in \code{\link[MatrixEQTL]{Matrix_eQTL_main}}.
If a file with this name exists it is overwritten, if \code{NULL} output is not saved
to file.}

\item{bootstrap_file_directory}{\code{character} or \code{NULL}. If not \code{NULL},
files will be saved in this directory storing local \emph{cis} associations for
the bootstrap eGene detection group (detection_bootstrapnumber.txt) and local
\emph{cis} associations the bootstrap left-out eGene effect size estimation
group (estimation_bootstrapnumber.txt). Estimation group files will only be saved
where signficant eGenes are also significant in the bootstrap detection group
(see Details). Corresponds to the \code{output_file_name.cis} argument in the
respective calls to \code{\link[MatrixEQTL]{Matrix_eQTL_main}}. Files in this
directory will be overwritten if they already exist.}

\item{cisDist}{\code{numeric}. Argument to \code{\link[MatrixEQTL]{Matrix_eQTL_main}}
controlling maximum distance SNP-gene pairs are considered local. The distance is
measured to the nearest end of the gene.}

\item{local_correction}{multiple testing correction method to use when
correcting p-values across all SNPs at each eGene. Must be a method
specified in \code{\link[stats]{p.adjust.methods}} or "qvalue" for
the \code{\link[qvalue]{qvalue}} package.}

\item{global_correction}{multiple testing correction method to use when
correcting p-values across all eGenes. Must be a method specified in
\code{\link[stats]{p.adjust.methods}} or "qvalue" for the
\code{\link[qvalue]{qvalue}} package.}

\item{bootstrap_eSNPs}{\code{character}. One of "discovery" or "top".
Controls which SNPs are used for effect size estimation in the bootstrap
procedure. If "discovery" effect sizes are estimated based on the top
eSNPs from the eGene detection analysis performed prior to the bootstrap
procedure. If "top", effect sizes are estimated based on the top eSNP
in each bootstrap (note: this makes the bootstrap procedure very slow;
see Details).}

\item{correction_type}{\code{character}. One of "shrinkage", "out_of_sample"
or "weighted". Determines which Winner's Curse correction method is
used (see Details).}

\item{collate_top_snps}{\code{logical}. If \code{'TRUE'} data on the most
frequent top SNP per eGene across all bootstraps is recorded (see Details).
Note setting this parameter to TRUE when \code{'bootstrap_eSNPs = "discovery"'}
will make the bootstrap procedure very slow since it requires computation
of all \emph{cis} associations at each eGene.}

\item{errorCovariance}{\code{numeric matrix} argument to \code{\link[MatrixEQTL]{Matrix_eQTL_main}}
specifying the error covariance.}

\item{useModel}{\code{integer} argument to \code{\link[MatrixEQTL]{Matrix_eQTL_main}}
specifying the type of model to fit between each SNP and gene. Should be one of
\code{\link[MatrixEQTL]{modelLINEAR}}, \code{\link[MatrixEQTL]{modelANOVA}}, or
\code{\link[MatrixEQTL]{modelLINEAR_CROSS}}.}
}
\value{
A \code{data.frame} (or \code{\link[data.table]{data.table}} if the
 user has the library loaded) containing the results for each significant eGene:
 \describe{
   \item{\code{'gene':}}{The eGene.}
   \item{\code{'top_snp':}}{The SNP with the smallest p-value for the eGene in the MatrixEQTL analysis.
     Multiple SNPs separated by a "/" indicate multiple SNPs in perfect LD with identical
     test statistics, effect sizes, and p-values.}
   \item{\code{'statistic':}}{The test statistic for the \code{top_snp}-\code{gene} pair.}
   \item{\code{'nominal_beta':}}{The eQTL effect size for the \code{top_snp}-\code{gene} pair in the
     MatrixEQTL analysis.}
   \item{\code{'nominal_pval':}}{The p-value for the \code{top_snp}-\code{gene} pair from the MatrixEQTL
     analysis prior to multiple testing correction.}
   \item{\code{'corrected_pval':}}{The hierarchical multiple-testing adjusted p-value for the
   \code{top_snp}-\code{gene} pair}
   \item{\code{'winners_curse':}}{The amount of effect size overestimation determined by the
     bootstrap analysis (See Details).}
   \item{\code{'corrected_beta':}}{The eGene effect size after adjustment for the \code{winners_curse}.}
   \item{\code{'correction_boots':}}{The number of bootstraps that contributed to the estimation of
     the \code{winners_curse}, \emph{i.e.} the number of bootstraps in which the
     \code{top_snp}-\code{gene} pair was significant (see details)}
   \item{\code{'best_boot_eSNP':}}{The SNP or SNPs that were most frequently the top eSNP in
     the bootstrap procedure (see details).}
   \item{\code{'prop_top_eSNP':}}{The proportion of bootstraps in which the \code{'best_boot_eSNP'}
     was the top eSNP (see details).}
 }
}
\description{
Performs cis-QTL mapping using MatrixEQTL then performs a bootstrap
analysis to obtain accurate effect size estimates for eGenes by
correcting for the "Winner's Curse".
}
\details{
\subsection{EGene detection:}{
 Detection of significant eGenes is performed using a hieararchical
 multiple testing correction procedure. At each gene, local multiple
 testing adjustment is performed using the method specified by the
 \code{'local_correction'} argument (Bonferroni correction by default)
 to correct for the number of \emph{cis}-SNPs tested at each gene.
 Global multiple testing adjustment is subsequently performed across
 all genes by taking the minimum locally corrected p-value at each
 gene, then performing the \code{'global_correction'} adjustment across
 all genes (FDR correction by default). These default settings best
 control the eGene false discovery rate without sacrificing
 sensitivity (see citation).
 }
 \subsection{Winner's Curse correction:}{
 EQTL effect sizes of significant eGenes are typically overestimated
 when compared to replication datasets due to the selection of the
 best \emph{cis}-SNP at each gene. \code{BootstrapEQTL} removes this
 overestimation by performing a bootstrap procedure after significant
 eGene discovery by \code{\link{MatrixEQTL}}.

 Three Winner's Curse correction methods are available: the Shrinkage
 method, the Out of Sample method, and the Weighted Estimator method.
 All three methods work on the same basic principle of performing
 repeated sample bootstrapping to partition the dataset into two
 groups: an eQTL detection group comprising study samples select via
 random sampling with replacement, and an eQTL effect size estimation
 group comprising the remaining samples not selected via the random
 sampling.

 The \strong{shrinkage method} ("shrinkage" in
 \code{'correction_type'}) corrects for the winners curse by
 measuring the average difference between the eGene effect size in
 the bootstrap detection group and the bootstrap estimation group,
 then subtracting this difference from the naive eQTL effect size
 estimate obtained from the eGene detection analysis prior to the
 bootstrap procedure.

 The \strong{out of sample method} ("out_of_sample" in
 \code{'correction_type'}) corrects for the winners curse by taking
 the average eGene effect size across bootstrap estimation groups as
 an unbiased effect size estimate.

 The \strong{weighted estimator method} ("weighted" in
 \code{'correction_type'}) corrects for the winners curse by taking a
 weighted average of the naive estimate of the effect size and the
 average of eGene effect sizes across the bootstrap estimation
 groups: \eqn{0.368 * naive_estimate + 0.632 *
 mean(bootstrap_estimation_group_effect_sizes)}.

 In all three methods bootstrap group effect sizes only contribute to
 the winner's curse correction if the corresponding eGene is
 significant in the bootstrap detection group. Two approaches are
 provided for measuring eGene significance and effect sizes in the
 bootstrap groups: (1) using the SNP in each bootstrap detection group
 with the smallest p-value ("top" in \code{'bootstrap_eSNPs'}), and
 (2) using the SNP with the smallest p-value in the eGene detection
 analysis performed prior to the bootstrap procedure ("discovery" in
 \code{'bootstrap_eSNPs'}). In both cases bootstraps are only
 considered in the final winner's curse calculation where the gene-SNP
 assocation is significant in the bootstrap detection group (see next
 section).

 The default settings, \code{'correction_type = "shrinkage"'} and
 \code{'bootstrap_eSNPs = "discovery"'} provided the most accurate
 corrected effect sizes in our simulation study (see citation).

 Note that use of \code{'bootstrap_eSNPs = "top"'} will
 dramatically slow down the bootstrap procedure as associaitons will
 need to be calculated between all \emph{cis}-SNPs and significant
 eGenes rather than just between the significant eGenes and their
 lead SNPs.

 Note that SNP-gene pairs may not remain significant in all
 bootstraps, so the effective number of bootstraps used obtain the
 "Winner's Curse" estimate will typically be lower than the number of
 bootstraps specified in \code{'n_bootstraps'}. The number of
 bootstraps that were significant for each eGene is reported in the
 \code{'correction_boots'} column of the returned table.
 }
 \subsection{Bootstrap inclusion}{
 The bootstrap significance threshold that determines the inclusion
 of any bootstrap to the final winners curse calculation of each
 gene-SNP pair is separate to the multiple testing correction
 procedure used for eGene discovery. It is calibrated to lead to
 optimal bootstrap inclusion for the most accurate winner's curse
 correction rather than for controlling eGene FDR.

 Gene-SNP associations at each bootstrap are bonferroni adjusted for
 the number of \emph{cis}-SNPs at each gene, and considered
 significant if they are smaller than the bootstrap significance
 threshold. The bootstrap significance threshold is determined prior
 to the bootstrap procedure. It is the local-Bonferroni threshold
 corresponding to the global multiple testing correction threshold
 of 0.05 in the eGene discovery analysis.
 }
 \subsection{Collating the most frequent top bootstrap SNP}{
 In addition to correcting for the Winner's Curse, the bootstrap
 procedure can also collects statistics about which SNP(s) are the top
 SNP for each tested eGene in the bootstrapped detection group. This
 can differ from the top SNP from the initial eGene detection
 analysis in any given bootstrap. The \code{'best_boot_eSNP'} column
 reports the SNP that was the top SNP most frequently across the
 bootstrap procedure. There is some evidence from simulation studies
 that this is more likely to be the causal SNP than the
 \code{'top_snp'} from the eGene detection analysis (see citation),
 although the \code{'top_snp'} and \code{'best_boot_eSNP'} are
 typically the same.

 Multiple SNPs may be reported in the \code{'best_boot_eSNP'} column.
 SNPs separated by a "/" indicate multiple SNPs in perfect linkage
 disequilbrium (LD), \emph{i.e.} with identical test
 statistics/effect sizes/p-values in all bootstraps, while SNPs
 separated by a ";" indicate multiple SNPs (or SNP groups in perfect
 LD) that were the top SNP(s) with equal frequency across the
 bootstrap procedure (\emph{e.g.} "SNP1;SNP2/SNP3" could indicate
 SNP1 was the top SNP in 50\% of significant bootstraps and SNP2 and
 SNP3 are in perfect LD and were was the top SNP in the other 50\% of
 significant bootstraps). The \code{'prop_top_eSNP'} column reports
 the proportion of significant bootstraps for which the
 \code{'best_boot_eSNP'} was the top SNP. A low
 \code{'prop_top_eSNP'} (\emph{e.g.} < 0.5) can indicate complex LD
 structure around the top eSNP or multiple causal eSNPs. Note that
 here the number of significant bootstraps may be higher than that
 reported in \code{'correction_boots'} if
 \code{'bootstrap_eSNPs="discovery"'}, as an eGene may be significant
 when performing global multiple testing correction across the top
 bootstrap detection group SNP but not when performing global
 multiple testing correction on the top SNP from the original eGene
 detection analysis.
 }
 \subsection{Bootstrap warning messages:}{
 It is possible for bootstrap analyses to fail due to the reduced
 sample sizes of the bootstrap detection and bootstrap estimation
 groups. For example, the bootstrap resampling may lead to an
 estimation group in which all individuals are homozygous for the
 eQTL SNP or are all the same age or sex. This causes the eQTL
 analysis to crash since the linear models cannot be fit. These
 errors are collated at the end of the bootstrap procedure, grouped
 together, and reported as a series of warnings detailing the number
 of bootstraps in which the error occurred as well as whether they
 occurred in the detection or estimation groups.
 }
}
\examples{
# Locations for example data from the MatrixEQTL package
base.dir = find.package('MatrixEQTL');
SNP_file_name = paste(base.dir, "/data/SNP.txt", sep="");
snps_location_file_name = paste(base.dir, "/data/snpsloc.txt", sep="");
expression_file_name = paste(base.dir, "/data/GE.txt", sep="");
gene_location_file_name = paste(base.dir, "/data/geneloc.txt", sep="");
covariates_file_name = paste(base.dir, "/data/Covariates.txt", sep="");

# Load the SNP data
snps = SlicedData$new();
snps$fileDelimiter = "\\t";      # the TAB character
snps$fileOmitCharacters = "NA"; # denote missing values;
snps$fileSkipRows = 1;          # one row of column labels
snps$fileSkipColumns = 1;       # one column of row labels
snps$fileSliceSize = 2000;      # read file in slices of 2,000 rows
snps$LoadFile(SNP_file_name);

# Load the data detailing the position of each SNP
snpspos = read.table(snps_location_file_name, header = TRUE, stringsAsFactors = FALSE);

# Load the gene expression data
gene = SlicedData$new();
gene$fileDelimiter = "\\t";      # the TAB character
gene$fileOmitCharacters = "NA"; # denote missing values;
gene$fileSkipRows = 1;          # one row of column labels
gene$fileSkipColumns = 1;       # one column of row labels
gene$fileSliceSize = 2000;      # read file in slices of 2,000 rows
gene$LoadFile(expression_file_name);

# Load the data detailing the position of each gene
genepos = read.table(gene_location_file_name, header = TRUE, stringsAsFactors = FALSE);

# Load the covariates data
cvrt = SlicedData$new();
cvrt$fileDelimiter = "\\t";      # the TAB character
cvrt$fileOmitCharacters = "NA"; # denote missing values;
cvrt$fileSkipRows = 1;          # one row of column labels
cvrt$fileSkipColumns = 1;       # one column of row labels
if(length(covariates_file_name)>0) {
  cvrt$LoadFile(covariates_file_name);
}

# Run the BootstrapQTL analysis
eGenes <- BootstrapQTL(snps, gene, snpspos, genepos, cvrt, n_bootstraps=10, n_cores=2)

}
